Move gosplit -> internal/gosplit
