+++ internal/gosplit/errors.go
