+++ internal/gosplit/gosplit.go
