+++ internal/gosplit/safeint.go
