+++ internal/gosplit/safeint_test.go
