+++ internal/gosplit/testdata/TestByBytesGolden.golden
