+++ internal/gosplit/testdata/example.txt
